scripts/requirements-cowork.txt
requests>=2.31.0
google-auth>=2.27.0
google-api-python-client>=2.118.0
